use setuptools to build portable executable.
