+++ mcsn_supervisor/.gitignore
@@ -1,2 +1,5 @@
 .venv/
 __pycache__/
+build/
+dist/
+*.egg-info
