--- mcsn_supervisor/mcsn.py
@@ -1,4 +1,0 @@
-from mcsn import mcsn_cli
-
-if __name__ == "__main__":
-    mcsn_cli()
